requirements.txt
streamlit
face_recognition
playwright
requests
fpdf2
Pillow
numpy